Simplified implementation and added better specs.
Don't rely on too many options for is :rateable.
Instead, use DataMapper property presence to decide
on which was previously decided via different options
to is :rateable. This ideally makes usage of
dm-is-rateable more consistent with what seems
to be a DataMapper convention: List all model properties
within the class body instead of automagically creating
them. This makes perfect sense to me, be it a convention
or not.